requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
transformers==4.35.0
torch==2.2.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.0
openai==1.1.0
